requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pymongo>=4.0
python-dotenv>=1.0
pydantic>=1.10
